+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED](https://github.com/quartiq/idsp/compare/v0.20.0..HEAD) - DATE
+
+### Changed
+
+* `Pid`: use `Units` to scale and `Build` trait to convert to `BiquadClamp`
+
 ## [0.20.0](https://github.com/quartiq/idsp/compare/v0.19.0..v0.20.0) - 2026-01-13
 
 ### Changed